feat(widget): trim CDN/IIFE bundle by excluding dev-only helpers
This update introduces a new `index-core.ts` file that separates the core API from development-only utilities, specifically `generateCodeSnippet` and `createDemoCarousel`, which are now excluded from the IIFE/CDN build. This change reduces the bundle size without affecting the functionality for npm consumers, who can still access these helpers through the `index.ts` barrel. The global `window.AgentWidget` no longer exposes these tools, ensuring a cleaner production environment.
